// src/app/model/signup.model.ts
export class RegisterSetupOrganisation {
    gemsOrgId: number;
    gemsOrgName: string;
    gemsOrgCode: string;
    gemsEmpCount: string;
    gemsOrgPhone: string;
    gemsOrgEmail: string;
    gemsOrgAddress: string;
    gemsOrgCity: string;
    gemsOrgState: string;
    adOrgZipCode: string;
    gemsOrgAdminName: string;
    gemsOrgSetupStatus: boolean;
  }
  